drop rcu_nocbs from cmdline
This option has been linked to a kernel deadlock we've experienced in
some of our clusters during VERY heavy I/O workloads. A testcase and
explanation will be sent to the RCU maintainers, but for now we're
just disabling it. It's possible the bug is not in RCU but elsewhere
in the kernel and this option just makes it more likely we hit the
buggy code path.

For safety, we'll remove it for now.